requirements.txt
# Core Dependencies
manim==0.18.0
groq==0.4.2
TTS
pyyaml==6.0.1
ffmpeg-python==0.2.0

# LLM & AI
requests==2.31.0

# Image Processing
pillow==10.2.0

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1

# Optional: Local LLM support
# ollama  # Uncomment if using Ollama Python SDK
